Hmm, now that works, lets hope Int64 is enough as summonerId

+++ Client/HextechFriendsClient/League/Protocol/CurrentSummoner.cs
@@ -8,7 +8,7 @@ namespace HextechFriendsClient.League.Protocol
 {
     public class CurrentSummoner
     {
-        public int accountId { get; set; }
+        public Int64 accountId { get; set; }
         public string displayName { get; set; }
         public string internalName { get; set; }
         public string lastSeasonHighestRank { get; set; }
@@ -16,7 +16,7 @@ namespace HextechFriendsClient.League.Protocol
         public int profileIconId { get; set; }
         public string puuid { get; set; }
         public RerollPoints rerollPoints { get; set; }
-        public int summonerId { get; set; }
+        public Int64 summonerId { get; set; }
         public int summonerLevel { get; set; }
         public int xpSinceLastLevel { get; set; }
         public int xpUntilNextLevel { get; set; }
